[feat](installment): add a InstallmentCalculator view
